+++ app/src/main/kotlin/com/sergebakharev/hnplus/model/HNConfirmForm.kt
@@ -1,0 +1,6 @@
+package com.sergebakharev.hnplus.model
+
+class HNConfirmForm(
+    val actionUrl: String,
+    val fields: Map<String, String>
+)